Add `libpak-tools package bundle` command which builds and packages a buildpack
This command runs the build workflow for libcnb & libpak based buildpacks. It both compiles (`libpak-tools package compile`) and also packages the compiled buildpack into a buildpack image which is loaded into the Docker daemon.

It does not presently work for non-libcnb/libpak buildpacks because they have a slightly different build system, and specifically the way that the version number is set differs. This can be worked around, but requires a separate PR to libcnb to update the compilation command and how it handles setting the version number.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>