Adds `build-jvm-vendors` command for building the JVM Vendors buildpack
This command can be used to build a single JVM Vendors buildpack with all of the vendors, it can be used to build individual buildpacks one per vendor, or it can even be used to build combinations with a subset of JVM vendors.

It is designed to be run locally, so one can easily package up JVM vendor buildpacks manually, while also still being suitable for use in CI. When used in CI, it should be the same commands as run locally.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>